Remove unneeded STL language operations
Removes ID_xor and ID_if from the list of possible options for
bool_synth_encoding. With ID_notequal there exists an alternative (which
is equivalent for bools) to ID_xor, hence one of them is not needed.
Regarding the ternary operator: While it was tempting to integrate it at
first because of the S and R operations in STL, reality shows that
converting the ternary operator to STL would need far more instructions
than expected.