Split student project into multiple tests
When synthesising multiple expressions of different complexity, there is
a high chance that overly complex solutions are retrieved for the
simpler expressions. Expressions which are independent from each other
should thus be synthesised separately.

Additionally, this commit adds non-deterministic booleans to the tests,
which greatly decreases the size of the specification.