Rework bool_synth_encoding strategy
Changes the way how options, constraints and solutions of boolean
expressions are generated. Each instruction now represents one line of
code in STL, thus the options of each instruction now include STL IDs
and one operand at most. The new implementation increases the synthesis
speed and quality of all samples considerably since the final size of
the STL program is taken into account. Expressions, which are simple in
theory but become overly complex when converted to STL, won't be
generated for a size too small for the STL program.

Limitations: This commit does not include nested boolean instructions.
The generated constraints are sometimes redundant, e.g. 'par0 && par0'
can occur, yielding to candidates that can not be properly translated to
STL after simplification. This does not cause any harm since for each
invalid combination there are valid and simpler alternatives, thus an
invalid combination should never be the result of a synthesis process.
Still they need to be removed in future work to avoid problems when
expanding the instruction set and to increase the execution speed even
more.